フォルダ名とスクリプト名を指定できるようにしました
