--- Test/Assets/Scripts/Versioned/Bullet1_backup_20250421_230405/Bullet1_backup_20250421_230405.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4e886839c8bd0774fac46d6ec1705102
