+++ Test/Assets/Scripts/Versioned/Bullet_Ver01/Bullet_Ver01.cs
@@ -3,8 +3,7 @@ using System.Collections.Generic;
 //using System.Numerics;
 using UnityEngine;
 using Scripts;
-namespace Backup_Bullet_backup_20250421_230037 
-
+namespace Backup_Bullet_Ver01 
 {
     public class Bullet : MonoBehaviour
     {
@@ -26,9 +25,6 @@ namespace Backup_Bullet_backup_20250421_230037
 
         private UnityEngine.Vector3 SavePower;
 
-        [SerializeField] private float reflectionCooldown = 1.0f; // 反射できるようになるまでの秒数
-        private float spawnTime; // 発射された時間
-
         private void Start()
         {
             Power *= PowerDirection;
@@ -37,7 +33,6 @@ namespace Backup_Bullet_backup_20250421_230037
             Debug.Log(meshRendererChild.name);
             reflectionCount = 0;
             cameraAreaManager = GameObject.FindObjectOfType<CameraAreaManager>();
-            spawnTime = Time.time; // 現在の時間を記録
         }
 
         void Update()
@@ -65,13 +60,11 @@ namespace Backup_Bullet_backup_20250421_230037
                     pos.y = cameraAreaManager.DownMax;
 
                 transform.position = pos;
-                spawnTime = 1;
             }
         }
 
         private void OnTriggerEnter2D(Collider2D collision)
         {
-
             if (collision.gameObject.tag == "Ground" || (collision.gameObject.tag == "Player" && !isAttack))
             {
                 if (collision.TryGetComponent<PlayerStatus>(out PlayerStatus status))
@@ -86,14 +79,24 @@ namespace Backup_Bullet_backup_20250421_230037
                 Destroy(this.gameObject);
             }
 
-            if (collision.gameObject.tag == "Attack" && !destroyed && Time.time - spawnTime >= reflectionCooldown)
+            if (collision.gameObject.tag == "Attack" && !destroyed)
             {
                 player.isMove = false;
                 player.Arrow.SetActive(true);
                 isAttack = true;
+                Time.timeScale = 0.2f;
+                Power = UnityEngine.Vector3.zero;
+                Invoke("Attack", 0.3f);
+            }
+
+            if (collision.gameObject.tag == "QuickAttack" && !destroyed)
+            {
+                player.isMove = false;
+                isAttack = true;
+                Time.timeScale = 0.2f;
                 SavePower = -Power;
                 Power = UnityEngine.Vector3.zero;
-                Invoke("Attack", 0.1f);
+                Invoke("QuickAttack", 0.1f);
             }
         }
 
@@ -111,15 +114,13 @@ namespace Backup_Bullet_backup_20250421_230037
             player.Arrow.SetActive(false);
             player.isMove = true;
             Invoke("AttckFalse", 0.2f);
-
             PowerDirection *= 1.25f;
             if (PowerDirection < 0)
                 PowerDirection *= -1;
-
             float Angle = Mathf.Atan2(player.InputMove.y, player.InputMove.x);
             UnityEngine.Vector3 direction = new UnityEngine.Vector3(Mathf.Cos(Angle), Mathf.Sin(Angle), 0);
             Power = direction * PowerDirection * 10f;
-
+            Time.timeScale = 1f;
             player.PlayReflectionSound();
         }
 
@@ -159,7 +160,6 @@ namespace Backup_Bullet_backup_20250421_230037
         {
             player.PlayReflectionSound();
         }
-
         void AttckFalse()
         {
             isAttack = false;
