+++ Test/Assets/Scripts/Versioned/Bullet_Ver01/Bullet_Ver01_memo.txt.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 8174ae3f7c60d63489d66860b41a099e
-folderAsset: yes
-DefaultImporter:
+guid: 1b07bab4788373d429bbadd3b64d2196
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
