+++ Test/Assets/Scripts/Versioned/Bullet_Ver02.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a102f3ad24702664b9c65ea6fd667571
-TextScriptImporter:
+guid: 3eec6734aacc8aa468b59eb37727b52b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
