+++ Test/Assets/Scripts/Versioned/Bullet_Ver02/Bullet_Ver02.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 //using System.Numerics;
 using UnityEngine;
 using Scripts;
-namespace Backup_Bullet1_backup_20250421_230405
+namespace Backup_Bullet_Ver02 
 {
     public class Bullet : MonoBehaviour
     {
@@ -25,6 +25,9 @@ namespace Backup_Bullet1_backup_20250421_230405
 
         private UnityEngine.Vector3 SavePower;
 
+        [SerializeField] private float reflectionCooldown = 1.0f; // 反射できるようになるまでの秒数
+        private float spawnTime; // 発射された時間
+
         private void Start()
         {
             Power *= PowerDirection;
@@ -33,6 +36,7 @@ namespace Backup_Bullet1_backup_20250421_230405
             Debug.Log(meshRendererChild.name);
             reflectionCount = 0;
             cameraAreaManager = GameObject.FindObjectOfType<CameraAreaManager>();
+            spawnTime = Time.time; // 現在の時間を記録
         }
 
         void Update()
@@ -60,11 +64,13 @@ namespace Backup_Bullet1_backup_20250421_230405
                     pos.y = cameraAreaManager.DownMax;
 
                 transform.position = pos;
+                spawnTime = 1;
             }
         }
 
         private void OnTriggerEnter2D(Collider2D collision)
         {
+
             if (collision.gameObject.tag == "Ground" || (collision.gameObject.tag == "Player" && !isAttack))
             {
                 if (collision.TryGetComponent<PlayerStatus>(out PlayerStatus status))
@@ -79,24 +85,14 @@ namespace Backup_Bullet1_backup_20250421_230405
                 Destroy(this.gameObject);
             }
 
-            if (collision.gameObject.tag == "Attack" && !destroyed)
+            if (collision.gameObject.tag == "Attack" && !destroyed && Time.time - spawnTime >= reflectionCooldown)
             {
                 player.isMove = false;
                 player.Arrow.SetActive(true);
                 isAttack = true;
-                Time.timeScale = 0.2f;
-                Power = UnityEngine.Vector3.zero;
-                Invoke("Attack", 0.3f);
-            }
-
-            if (collision.gameObject.tag == "QuickAttack" && !destroyed)
-            {
-                player.isMove = false;
-                isAttack = true;
-                Time.timeScale = 0.2f;
                 SavePower = -Power;
                 Power = UnityEngine.Vector3.zero;
-                Invoke("QuickAttack", 0.1f);
+                Invoke("Attack", 0.1f);
             }
         }
 
@@ -114,13 +110,15 @@ namespace Backup_Bullet1_backup_20250421_230405
             player.Arrow.SetActive(false);
             player.isMove = true;
             Invoke("AttckFalse", 0.2f);
+
             PowerDirection *= 1.25f;
             if (PowerDirection < 0)
                 PowerDirection *= -1;
+
             float Angle = Mathf.Atan2(player.InputMove.y, player.InputMove.x);
             UnityEngine.Vector3 direction = new UnityEngine.Vector3(Mathf.Cos(Angle), Mathf.Sin(Angle), 0);
             Power = direction * PowerDirection * 10f;
-            Time.timeScale = 1f;
+
             player.PlayReflectionSound();
         }
 
@@ -160,6 +158,7 @@ namespace Backup_Bullet1_backup_20250421_230405
         {
             player.PlayReflectionSound();
         }
+
         void AttckFalse()
         {
             isAttack = false;
