+++ Test/Assets/Scripts/Versioned/Bullet_Ver02/Bullet_Ver02_memo.txt.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 4668d932521aa9e49a9015d55efdaccb
-folderAsset: yes
-DefaultImporter:
+guid: afe753aa6aa56c04b8980cd3f0d6b36d
+TextScriptImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
