--- Test/Assets/Scripts/Versioned/Bullet_backup_20250421_230037/Bullet_backup_20250421_230037.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 57272c7c54d0c714992731cd001fcb5f
