--- Test/Assets/Scripts/Versioned/Player1_backup_20250421_230456/Player1_backup_20250421_230456.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: a89384a0681e7ea4ba3ce2e136460c3c
