--- Test/Assets/Scripts/Versioned/Player1_backup_20250421_230456/memo.txt
@@ -1,1 +1,0 @@
-ver0.1　変更前
