+++ Test/Assets/Scripts/Versioned/Player_Ver01.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: a0f54e17b06937e4ca24569ad4a0ebab
-TextScriptImporter:
+guid: 7571dda86968a2d499d3ef47a631449b
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
