+++ Test/Assets/Scripts/Versioned/Player_Ver01/Player_Ver01.cs
@@ -1,24 +1,14 @@
 using System;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.Serialization;
 using UnityEngine.UI;
-using UnityEngine.VFX;
 using Scripts;
-namespace Backup_Player_backup_20250421_225603 
+namespace Backup_Player_Ver01 
 {
-    [Serializable]
-    public class VFXEntry
-    {
-        public string name;
-        public GameObject vfxObject;
-    }
-
     public class Player : MonoBehaviour
     {
         public static Player Instance;
-
         [SerializeField] private PlayerInput MoveAction;
         [SerializeField] private float MoveSpeed;
         public Vector2 InputMove = Vector2.zero;
@@ -40,14 +30,6 @@ namespace Backup_Player_backup_20250421_225603
         [SerializeField] [JapaneseLabel("2回目のジャンプまでのクールタイム")]
         private float jumpCooldown = 0.2f;
 
-        [FormerlySerializedAs("limitSpeed")] [SerializeField]
-        private List<VFXEntry> vfxEntries = new();
-
-        [SerializeField] private CameraAreaManager cameraAreaManager;
-        [SerializeField] private MapManager mapManager;
-        [SerializeField] private SceneButtonManager sceneButtonManager;
-        [SerializeField] private float maxFallSpeed = 20f;
-
         private AudioSource audioSource;
         [NonSerialized] public float BulletTime;
         [NonSerialized] public int direction = 1;
@@ -58,9 +40,11 @@ namespace Backup_Player_backup_20250421_225603
         private float lastJumpTime; // 最後にジャンプした時間
         private Rigidbody2D rb;
         private float startY;
-        private readonly Dictionary<string, GameObject> vfxDictionary = new();
+        [SerializeField] CameraAreaManager cameraAreaManager;
+        [SerializeField] MapManager mapManager;
 
-        private bool isArrowActive = false;
+        [FormerlySerializedAs("limitSpeed")] [SerializeField]
+        private float maxFallSpeed = 20f;
 
         private void Awake()
         {
@@ -76,25 +60,20 @@ namespace Backup_Player_backup_20250421_225603
             MoveAction.actions["Move"].canceled += OnMove;
             MoveAction.actions["Jump"].started += OnJump;
             MoveAction.actions["Shot"].started += OnShot;
-            MoveAction.actions["Attack"].started += OnAttackStart;
-            MoveAction.actions["Attack"].canceled += OnAttackEnd;
+            MoveAction.actions["Attack"].performed += OnAttack;
             MoveAction.actions["Jump"].canceled += OffJump;
-            // MoveAction.actions["QuickAttack"].performed += OnQuickAttack;
+            MoveAction.actions["QuickAttack"].performed += OnQuickAttack;
 
             rb = GetComponent<Rigidbody2D>();
             Arrow.SetActive(false);
             jumpCount = MaxJumpCount;
             audioSource = GetComponent<AudioSource>();
-            cameraAreaManager = FindObjectOfType<CameraAreaManager>();
-            mapManager = FindObjectOfType<MapManager>();
-            sceneButtonManager = FindObjectOfType<SceneButtonManager>();
 
-            foreach (var entry in vfxEntries)
-                if (!vfxDictionary.ContainsKey(entry.name))
-                    vfxDictionary.Add(entry.name, entry.vfxObject);
+            cameraAreaManager = GameObject.FindObjectOfType<CameraAreaManager>();
+            mapManager = GameObject.FindObjectOfType<MapManager>();
         }
 
-        private void FixedUpdate()
+        private void Update()
         {
             BulletUI.fillAmount = (MaxBulletTime - BulletTime) / MaxBulletTime;
 
@@ -106,7 +85,7 @@ namespace Backup_Player_backup_20250421_225603
                 }
                 else
                 {
-                    var pos = transform.position;
+                    Vector3 pos = transform.position;
 
                     if (pos.x < cameraAreaManager.LeftMax)
                         pos.x = cameraAreaManager.RightMax;
@@ -115,20 +94,31 @@ namespace Backup_Player_backup_20250421_225603
 
                     if (pos.y < cameraAreaManager.DownMax)
                     {
-                        if (mapManager.CanLoop(pos, MapManager.Side.down))
+                        if (mapManager.CanLoop(pos, MapManager.Side.down) == true)
+                        {
                             pos.y = cameraAreaManager.UpMax;
+                        }
                         else
+                        {
                             pos.y = cameraAreaManager.DownMax;
+                        }
 
+                        // Debug.Log(rb.linearVelocity);
                         if (rb.linearVelocity.y < maxFallSpeed * -1)
+                        {
                             rb.linearVelocity = new Vector2(rb.linearVelocity.x, maxFallSpeed * -1);
+                        }
                     }
                     else if (pos.y > cameraAreaManager.UpMax)
                     {
-                        if (mapManager.CanLoop(pos, MapManager.Side.up))
+                        if (mapManager.CanLoop(pos, MapManager.Side.up) == true)
+                        {
                             pos.y = cameraAreaManager.DownMax;
+                        }
                         else
+                        {
                             pos.y = cameraAreaManager.UpMax;
+                        }
                     }
 
                     transform.position = pos;
@@ -139,7 +129,6 @@ namespace Backup_Player_backup_20250421_225603
                 BulletTime -= Time.deltaTime;
             if (!isMove)
                 return;
-
             if (InputMove.x < 0)
             {
                 transform.position += new Vector3(MoveSpeed * InputMove.x, 0, 0) * Time.deltaTime;
@@ -158,30 +147,24 @@ namespace Backup_Player_backup_20250421_225603
 
         private void OnCollisionEnter2D(Collision2D collision)
         {
-            if (collision.gameObject.CompareTag("Ground"))
+            if (collision.gameObject.tag == "Ground")
                 jumpCount = MaxJumpCount;
         }
 
         public void OnMove(InputAction.CallbackContext context)
         {
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
+            animator.SetBool("isMove", true);
             InputMove = context.ReadValue<Vector2>();
-            isArrowActive = InputMove != Vector2.zero;
-
-            animator.SetBool("isMove", isArrowActive);
-
-            if (isArrowActive)
-            {
-                var angle = Mathf.Atan2(InputMove.y, InputMove.x) * Mathf.Rad2Deg;
-                Arrow.transform.rotation = Quaternion.Euler(0f, 0f, angle);
-            }
+            if (InputMove != Vector2.zero)
+                animator.SetBool("isMove", true);
+            else
+                animator.SetBool("isMove", false);
+            var Angle = Mathf.Atan2(InputMove.y, InputMove.x) * Mathf.Rad2Deg;
+            Arrow.transform.rotation = Quaternion.Euler(0f, 0f, Angle);
         }
 
         public void OnJump(InputAction.CallbackContext context)
         {
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
             if (jumpCount > 0 && Time.time - lastJumpTime >= jumpCooldown)
             {
                 rb.linearVelocity = new Vector2(rb.linearVelocity.x, jumpPower);
@@ -193,18 +176,18 @@ namespace Backup_Player_backup_20250421_225603
 
         public void OffJump(InputAction.CallbackContext context)
         {
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
             isJump = false;
             animator.SetBool("isJump", false);
         }
 
         public void OnShot(InputAction.CallbackContext context)
         {
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
             if (BulletTime <= 0)
             {
+                // audioSource.PlayOneShot(ShotSound);
+                // var bullets = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
+                // var bullet = bullets.GetComponent<Bullet>();
+                // bullet.PowerDirection = direction;
                 BulletTime = MaxBulletTime;
                 animator.SetTrigger("isShot");
             }
@@ -212,51 +195,31 @@ namespace Backup_Player_backup_20250421_225603
 
         public void Shot()
         {
+            Debug.Log("aaaa");
             audioSource.PlayOneShot(ShotSound);
-            var bulletObj = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
-            var bullet = bulletObj.GetComponent<Bullet>();
+            var bullets = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
+            var bullet = bullets.GetComponent<Bullet>();
             bullet.PowerDirection = direction;
         }
 
-        /// <summary>
-        ///     攻撃ボタン押下時の処理：攻撃判定と矢印を表示し、移動を無効化する
-        /// </summary>
-        public void OnAttackStart(InputAction.CallbackContext context)
+        public void OnAttack(InputAction.CallbackContext context)
         {
-            AttackCollision.SetActive(true);
-            Arrow.SetActive(true);
-            isMove = false;
-            // 必要に応じてアニメーションも再生
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
-            AttackCollision.SetActive(true);
-            Arrow.SetActive(true);
-            isMove = false;
-
+            AttackCollision.gameObject.SetActive(true);
+            Invoke("AttackFinish", 0.3f);
             animator.SetTrigger("isAttack");
         }
 
-        /// <summary>
-        ///     攻撃ボタン離し時の処理：攻撃判定と矢印を非表示にし、移動を再開する
-        /// </summary>
-        public void OnAttackEnd(InputAction.CallbackContext context)
+        public void OnQuickAttack(InputAction.CallbackContext context)
         {
-            AttackCollision.SetActive(false);
-            Arrow.SetActive(false);
-            isMove = true;
-
-            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
-
-            AttackCollision.SetActive(false);
-            Arrow.SetActive(false);
-            isMove = true;
+            QuickAttackCollision.gameObject.SetActive(true);
+            Invoke("AttackFinish", 0.3f);
             animator.SetTrigger("isAttack");
         }
 
         public void AttackFinish()
         {
-            AttackCollision.SetActive(false);
-            QuickAttackCollision.SetActive(false);
+            AttackCollision.gameObject.SetActive(false);
+            QuickAttackCollision.gameObject.SetActive(false);
         }
 
         public void PlayReflectionSound()
@@ -266,29 +229,7 @@ namespace Backup_Player_backup_20250421_225603
 
         public void PlayDamageSound()
         {
-            animator.Play("Damage");
             audioSource.PlayOneShot(DamageSound);
-        }
-
-        public void TriggerVFX(string vfxName)
-        {
-            if (vfxDictionary.TryGetValue(vfxName, out var vfxObject))
-            {
-                if (vfxObject.TryGetComponent<VisualEffect>(out var vfx))
-                    vfx.SendEvent("OnPlay"); // VFX のイベントを送信
-                else
-                    Debug.LogWarning($"指定されたVFXオブジェクト '{vfxName}' に VisualEffect コンポーネントがありません。");
-            }
-            else
-            {
-                Debug.LogWarning($"VFX '{vfxName}' が見つかりません。");
-            }
-        }
-
-        public Vector3 GetArrowDirection()
-        {
-            var angle = Arrow.transform.rotation.eulerAngles.z * Mathf.Deg2Rad;
-            return new Vector3(Mathf.Cos(angle), Mathf.Sin(angle), 0).normalized;
         }
     }
 }
