+++ Test/Assets/Scripts/Versioned/Player_Ver01/Player_Ver01_memo.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ddf3e9d7d88d86640a5330b1c2b44567
+guid: 1607a15a228d2874f935b654c73cdf7e
 TextScriptImporter:
   externalObjects: {}
   userData:
