+++ Test/Assets/Scripts/Versioned/Player_Ver02/Player_Ver02.cs
@@ -1,14 +1,24 @@
 using System;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.InputSystem;
 using UnityEngine.Serialization;
 using UnityEngine.UI;
+using UnityEngine.VFX;
 using Scripts;
-namespace Backup_Player1_backup_20250421_230456
+namespace Backup_Player_Ver02 
 {
+    [Serializable]
+    public class VFXEntry
+    {
+        public string name;
+        public GameObject vfxObject;
+    }
+
     public class Player : MonoBehaviour
     {
         public static Player Instance;
+
         [SerializeField] private PlayerInput MoveAction;
         [SerializeField] private float MoveSpeed;
         public Vector2 InputMove = Vector2.zero;
@@ -30,6 +40,14 @@ namespace Backup_Player1_backup_20250421_230456
         [SerializeField] [JapaneseLabel("2回目のジャンプまでのクールタイム")]
         private float jumpCooldown = 0.2f;
 
+        [FormerlySerializedAs("limitSpeed")] [SerializeField]
+        private List<VFXEntry> vfxEntries = new();
+
+        [SerializeField] private CameraAreaManager cameraAreaManager;
+        [SerializeField] private MapManager mapManager;
+        [SerializeField] private SceneButtonManager sceneButtonManager;
+        [SerializeField] private float maxFallSpeed = 20f;
+
         private AudioSource audioSource;
         [NonSerialized] public float BulletTime;
         [NonSerialized] public int direction = 1;
@@ -40,11 +58,9 @@ namespace Backup_Player1_backup_20250421_230456
         private float lastJumpTime; // 最後にジャンプした時間
         private Rigidbody2D rb;
         private float startY;
-        [SerializeField] CameraAreaManager cameraAreaManager;
-        [SerializeField] MapManager mapManager;
+        private readonly Dictionary<string, GameObject> vfxDictionary = new();
 
-        [FormerlySerializedAs("limitSpeed")] [SerializeField]
-        private float maxFallSpeed = 20f;
+        private bool isArrowActive = false;
 
         private void Awake()
         {
@@ -60,20 +76,25 @@ namespace Backup_Player1_backup_20250421_230456
             MoveAction.actions["Move"].canceled += OnMove;
             MoveAction.actions["Jump"].started += OnJump;
             MoveAction.actions["Shot"].started += OnShot;
-            MoveAction.actions["Attack"].performed += OnAttack;
+            MoveAction.actions["Attack"].started += OnAttackStart;
+            MoveAction.actions["Attack"].canceled += OnAttackEnd;
             MoveAction.actions["Jump"].canceled += OffJump;
-            MoveAction.actions["QuickAttack"].performed += OnQuickAttack;
+            // MoveAction.actions["QuickAttack"].performed += OnQuickAttack;
 
             rb = GetComponent<Rigidbody2D>();
             Arrow.SetActive(false);
             jumpCount = MaxJumpCount;
             audioSource = GetComponent<AudioSource>();
+            cameraAreaManager = FindObjectOfType<CameraAreaManager>();
+            mapManager = FindObjectOfType<MapManager>();
+            sceneButtonManager = FindObjectOfType<SceneButtonManager>();
 
-            cameraAreaManager = GameObject.FindObjectOfType<CameraAreaManager>();
-            mapManager = GameObject.FindObjectOfType<MapManager>();
+            foreach (var entry in vfxEntries)
+                if (!vfxDictionary.ContainsKey(entry.name))
+                    vfxDictionary.Add(entry.name, entry.vfxObject);
         }
 
-        private void Update()
+        private void FixedUpdate()
         {
             BulletUI.fillAmount = (MaxBulletTime - BulletTime) / MaxBulletTime;
 
@@ -85,7 +106,7 @@ namespace Backup_Player1_backup_20250421_230456
                 }
                 else
                 {
-                    Vector3 pos = transform.position;
+                    var pos = transform.position;
 
                     if (pos.x < cameraAreaManager.LeftMax)
                         pos.x = cameraAreaManager.RightMax;
@@ -94,31 +115,20 @@ namespace Backup_Player1_backup_20250421_230456
 
                     if (pos.y < cameraAreaManager.DownMax)
                     {
-                        if (mapManager.CanLoop(pos, MapManager.Side.down) == true)
-                        {
+                        if (mapManager.CanLoop(pos, MapManager.Side.down))
                             pos.y = cameraAreaManager.UpMax;
-                        }
                         else
-                        {
                             pos.y = cameraAreaManager.DownMax;
-                        }
 
-                        // Debug.Log(rb.linearVelocity);
                         if (rb.linearVelocity.y < maxFallSpeed * -1)
-                        {
                             rb.linearVelocity = new Vector2(rb.linearVelocity.x, maxFallSpeed * -1);
-                        }
                     }
                     else if (pos.y > cameraAreaManager.UpMax)
                     {
-                        if (mapManager.CanLoop(pos, MapManager.Side.up) == true)
-                        {
+                        if (mapManager.CanLoop(pos, MapManager.Side.up))
                             pos.y = cameraAreaManager.DownMax;
-                        }
                         else
-                        {
                             pos.y = cameraAreaManager.UpMax;
-                        }
                     }
 
                     transform.position = pos;
@@ -129,6 +139,7 @@ namespace Backup_Player1_backup_20250421_230456
                 BulletTime -= Time.deltaTime;
             if (!isMove)
                 return;
+
             if (InputMove.x < 0)
             {
                 transform.position += new Vector3(MoveSpeed * InputMove.x, 0, 0) * Time.deltaTime;
@@ -147,24 +158,30 @@ namespace Backup_Player1_backup_20250421_230456
 
         private void OnCollisionEnter2D(Collision2D collision)
         {
-            if (collision.gameObject.tag == "Ground")
+            if (collision.gameObject.CompareTag("Ground"))
                 jumpCount = MaxJumpCount;
         }
 
         public void OnMove(InputAction.CallbackContext context)
         {
-            animator.SetBool("isMove", true);
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
             InputMove = context.ReadValue<Vector2>();
-            if (InputMove != Vector2.zero)
-                animator.SetBool("isMove", true);
-            else
-                animator.SetBool("isMove", false);
-            var Angle = Mathf.Atan2(InputMove.y, InputMove.x) * Mathf.Rad2Deg;
-            Arrow.transform.rotation = Quaternion.Euler(0f, 0f, Angle);
+            isArrowActive = InputMove != Vector2.zero;
+
+            animator.SetBool("isMove", isArrowActive);
+
+            if (isArrowActive)
+            {
+                var angle = Mathf.Atan2(InputMove.y, InputMove.x) * Mathf.Rad2Deg;
+                Arrow.transform.rotation = Quaternion.Euler(0f, 0f, angle);
+            }
         }
 
         public void OnJump(InputAction.CallbackContext context)
         {
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
             if (jumpCount > 0 && Time.time - lastJumpTime >= jumpCooldown)
             {
                 rb.linearVelocity = new Vector2(rb.linearVelocity.x, jumpPower);
@@ -176,18 +193,18 @@ namespace Backup_Player1_backup_20250421_230456
 
         public void OffJump(InputAction.CallbackContext context)
         {
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
             isJump = false;
             animator.SetBool("isJump", false);
         }
 
         public void OnShot(InputAction.CallbackContext context)
         {
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
             if (BulletTime <= 0)
             {
-                // audioSource.PlayOneShot(ShotSound);
-                // var bullets = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
-                // var bullet = bullets.GetComponent<Bullet>();
-                // bullet.PowerDirection = direction;
                 BulletTime = MaxBulletTime;
                 animator.SetTrigger("isShot");
             }
@@ -195,31 +212,51 @@ namespace Backup_Player1_backup_20250421_230456
 
         public void Shot()
         {
-            Debug.Log("aaaa");
             audioSource.PlayOneShot(ShotSound);
-            var bullets = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
-            var bullet = bullets.GetComponent<Bullet>();
+            var bulletObj = Instantiate(Bullets, ShotPosition.transform.position, Quaternion.identity);
+            var bullet = bulletObj.GetComponent<Bullet>();
             bullet.PowerDirection = direction;
         }
 
-        public void OnAttack(InputAction.CallbackContext context)
+        /// <summary>
+        ///     攻撃ボタン押下時の処理：攻撃判定と矢印を表示し、移動を無効化する
+        /// </summary>
+        public void OnAttackStart(InputAction.CallbackContext context)
         {
-            AttackCollision.gameObject.SetActive(true);
-            Invoke("AttackFinish", 0.3f);
+            AttackCollision.SetActive(true);
+            Arrow.SetActive(true);
+            isMove = false;
+            // 必要に応じてアニメーションも再生
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
+            AttackCollision.SetActive(true);
+            Arrow.SetActive(true);
+            isMove = false;
+
             animator.SetTrigger("isAttack");
         }
 
-        public void OnQuickAttack(InputAction.CallbackContext context)
+        /// <summary>
+        ///     攻撃ボタン離し時の処理：攻撃判定と矢印を非表示にし、移動を再開する
+        /// </summary>
+        public void OnAttackEnd(InputAction.CallbackContext context)
         {
-            QuickAttackCollision.gameObject.SetActive(true);
-            Invoke("AttackFinish", 0.3f);
+            AttackCollision.SetActive(false);
+            Arrow.SetActive(false);
+            isMove = true;
+
+            if (sceneButtonManager.CurrentState != SceneButtonManager.State.Gameplay) return;
+
+            AttackCollision.SetActive(false);
+            Arrow.SetActive(false);
+            isMove = true;
             animator.SetTrigger("isAttack");
         }
 
         public void AttackFinish()
         {
-            AttackCollision.gameObject.SetActive(false);
-            QuickAttackCollision.gameObject.SetActive(false);
+            AttackCollision.SetActive(false);
+            QuickAttackCollision.SetActive(false);
         }
 
         public void PlayReflectionSound()
@@ -229,7 +266,29 @@ namespace Backup_Player1_backup_20250421_230456
 
         public void PlayDamageSound()
         {
+            animator.Play("Damage");
             audioSource.PlayOneShot(DamageSound);
+        }
+
+        public void TriggerVFX(string vfxName)
+        {
+            if (vfxDictionary.TryGetValue(vfxName, out var vfxObject))
+            {
+                if (vfxObject.TryGetComponent<VisualEffect>(out var vfx))
+                    vfx.SendEvent("OnPlay"); // VFX のイベントを送信
+                else
+                    Debug.LogWarning($"指定されたVFXオブジェクト '{vfxName}' に VisualEffect コンポーネントがありません。");
+            }
+            else
+            {
+                Debug.LogWarning($"VFX '{vfxName}' が見つかりません。");
+            }
+        }
+
+        public Vector3 GetArrowDirection()
+        {
+            var angle = Arrow.transform.rotation.eulerAngles.z * Mathf.Deg2Rad;
+            return new Vector3(Mathf.Cos(angle), Mathf.Sin(angle), 0).normalized;
         }
     }
 }
