+++ Test/Assets/Scripts/Versioned/Player_Ver02/Player_Ver02_memo.txt.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a892a37331345d9459204da61f8d5a37
+guid: be1ac7fb7c40d3e43a0fda4d824e04e0
 TextScriptImporter:
   externalObjects: {}
   userData:
