--- Test/Assets/Scripts/Versioned/Player_backup_20250421_225603/Player_backup_20250421_225603.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: eba201e6e67c2a84b803bf0129f9c836
